+++ src/Traits/ArrayStorageConfigurableTrait.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Traits;
+
+/**
+ * Trait ArrayStorageConfigurableTrait
+ * @package Php\Support\Traits
+ */
+trait ArrayStorageConfigurableTrait
+{
+    use ArrayStorage, ConfigurableTrait {
+        ArrayStorage::propertyExists insteadof ConfigurableTrait;
+    }
+
+    /**
+     * @param string $key
+     * @param $value
+     *
+     * @return bool
+     */
+    protected function setProp(string $key, $value): bool
+    {
+        $this->set($key, $value);
+
+        return true;
+    }
+
+}