+++ tests/traits/ConfigurableTest.php
@@ -47,4 +47,34 @@ final class ConfigurableTest extends TestCase
 
     }
 
+    public function testConfigurable2(): void
+    {
+        $cls = new ConfigurableTraitTestClass();
+
+        $this->assertNull($cls->prop);
+        $cls->configurable(['prop' => 'success', 'test' => 'fake', 'fn' => 123], false);
+
+        $this->assertEquals('success', $cls->prop);
+        $this->assertEquals(123, $cls->_fn);
+        $this->assertFalse(property_exists($cls, 'test'));
+    }
+
+}
+
+class ConfigurableTraitTestClass
+{
+    use \Php\Support\Traits\ConfigurableTrait;
+    public $prop;
+    public $_fn;
+
+
+    public function propertyExist(string $key): bool
+    {
+        return true;
+    }
+
+    public function setFn($val)
+    {
+        $this->_fn = $val;
+    }
 }